Deepfake-Detection-Tool-main/backend/requirements.txt
Flask
Flask-Cors
bcrypt
PyJWT
numpy
Pillow
tensorflow
